Bug 341 - suppress log trigger/deny when running in 'local' mode

+++ src/backend/slony1_funcs.sql
@@ -5520,13 +5520,20 @@ comment on function @NAMESPACE@.slon_node_health_check() is 'called when slon st
 create or replace function @NAMESPACE@.log_truncate () returns trigger as
 $$
 	declare
+		r_role text;
 		c_command text;
 		c_log integer;
 		c_node integer;
 		c_tabid integer;
 	begin
-        c_tabid := tg_argv[0];
-	    c_node := @NAMESPACE@.getLocalNodeId('_@CLUSTERNAME@');
+		-- Ignore this call if session_replication_role = 'local'
+		select into r_role setting
+			from pg_catalog.pg_settings where name = 'session_replication_role';
+		if r_role = 'local' then
+			return NULL;
+		end if;
+        	c_tabid := tg_argv[0];
+	    	c_node := @NAMESPACE@.getLocalNodeId('_@CLUSTERNAME@');
 		c_command := 'TRUNCATE TABLE ONLY "' || tab_nspname || '"."' ||
 				  tab_relname || '" CASCADE' 
 				  from @NAMESPACE@.sl_table where tab_id = c_tabid;
@@ -5548,7 +5555,16 @@ is 'trigger function run when a replicated table receives a TRUNCATE request';
 
 create or replace function @NAMESPACE@.deny_truncate () returns trigger as
 $$
+	declare
+		r_role text;
 	begin
+		-- Ignore this call if session_replication_role = 'local'
+		select into r_role setting
+			from pg_catalog.pg_settings where name = 'session_replication_role';
+		if r_role = 'local' then
+			return NULL;
+		end if;
+
 		raise exception 'truncation of replicated table forbidden on subscriber node';
     end
 $$ language plpgsql;
